TS-35786 jlink tests
